+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## flor 0.15.0  not yet released
 
+- Let "case" match regular expressions
+
 
 ## flor 0.14.0  released 2017-06-13